OC-9418 User service call failure results in Runtime not allowing users to login
