+++ web/pom.xml
@@ -86,7 +86,6 @@
         <dependency>
             <groupId>org.springframework.security</groupId>
             <artifactId>spring-security-jwt</artifactId>
-            <version>1.0.7.RELEASE</version>
         </dependency>
