+++ web/src/main/java/org/akaza/openclinica/service/StudyBuildService.java
@@ -9,6 +9,7 @@ import org.slf4j.LoggerFactory;
 import org.springframework.http.ResponseEntity;
 
 import javax.servlet.http.HttpServletRequest;
+import java.util.List;
 
 /**
  * Created by yogi on 11/10/16.
@@ -20,7 +21,7 @@ public interface StudyBuildService {
 
     boolean saveStudyEnvRoles(HttpServletRequest request, UserAccountBean ub) throws Exception;
 
-    ResponseEntity getUserRoles(HttpServletRequest request);
+    ResponseEntity<List<StudyEnvironmentRoleDTO>> getUserRoles(HttpServletRequest request);
 
     ResponseEntity getUserDetails(HttpServletRequest request);
     void updateStudyUsername(UserAccountBean ub, Auth0User user);
